3:17 PM added InputType enum class to ImageBase; 4:18 PM completed updating all conditional blocks to use input_type based switch case; removed uchar_input; 4:31 PM made input_type const; 4:47 PM made all of the remaining parameters in ImageBase const too; 4:56 PM made resx, resy and n_pts in StateSpaceModel const too; 5:16 PM updated SumOfAMs to handle the const variables in a more logical manner;

+++ SM/src/GridTracker.cc
@@ -316,7 +316,7 @@ void GridTracker<SSM>::backwardEstimation(){
 			curr_pts_masked.push_back(curr_pts[tracker_id]);
 		}
 	}
-	if(prev_pts_masked.size() < est_params.n_model_pts){
+	if(static_cast<int>(prev_pts_masked.size()) < est_params.n_model_pts){
 		for(int tracker_id = 0; tracker_id < n_trackers; ++tracker_id){
 			if(fb_err_mask[tracker_id]){ continue; }
 			prev_pts_masked.push_back(prev_pts[tracker_id]);
